--- enterprise/ha/src/main/resources/META-INF/services/org.neo4j.backup.BackupExtensionService
@@ -1,1 +1,0 @@
-org.neo4j.kernel.ha.backup.HaBackupProvider
