--- enterprise/ha/src/main/resources/META-INF/services/org.neo4j.kernel.impl.cache.CacheProvider
@@ -1,1 +1,0 @@
-org.neo4j.kernel.impl.cache.HighPerformanceCacheProvider
