--- enterprise/ha/src/main/resources/META-INF/services/org.neo4j.kernel.impl.locking.Locks$Factory
@@ -1,1 +1,0 @@
-org.neo4j.kernel.ha.lock.forseti.ForsetiLocksFactory
