+++ enterprise/kernel/src/main/java/org/neo4j/kernel/impl/enterprise/lock/forseti/ForsetiLockManager.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.ha.lock.forseti;
+package org.neo4j.kernel.impl.enterprise.lock.forseti;
 
 import java.util.Map;
 import java.util.Queue;
