+++ enterprise/kernel/src/main/resources/META-INF/services/org.neo4j.kernel.impl.locking.Locks$Factory
@@ -1,0 +1,1 @@
+org.neo4j.kernel.impl.enterprise.lock.forseti.ForsetiLocksFactory
